Promoting several consistently passing tests to the main suite
